Refactor everything
Refactor KlikBCA class and run.php file. KlikBCA file contains the core
of this scraper library and run.php file contains the usage example of
this library.

In this refactoring, we have so many improvements, especially for the
logout part. Previously, when we use this library, our KlikBCA session
will be left dangling and can't be used. This makes our account cannot
be accessed for 10 minutes (waiting for the dangling session to be
expired). After this refactoring, this issue is fixed by invoking the
logout endpoint in the KlikBCA class' destructor.

Thanks to Kurniawan for discussing this with me 4 years ago. See the
link below for the story.

Link: https://fb.com/2159278187435164
Signed-off-by: Ammar Faizi <ammarfaizi2@gnuweeb.org>