KlikBCA: Fix error handling in the `logout()` method
The error handling on curl error was wrong, it should set `$this->err`,
not `$err`. Setting in `$err` unused local variable doesn't mean
anything. This was a no-op then return false. What we should do is
setting `$this->err` then return false, so the caller can get the
error message via `getErr()` method.

Fixes: 688735105f7937260d7d2cfc8a8cc9288762ccfc ("KlikBCA: Fix return annotation comment in `setProxy()`")
Signed-off-by: Ammar Faizi <ammarfaizi2@gnuweeb.org>