KlikBCA: Introducing reuse session mechanism
To reduce the number of HTTP requests, create a mechanism that allows
the user to reuse the login session. It can avoid unnecessary login()
and logout() calls, thus improving the performance of the application.

To tell the library that the user wants to reuse the session, the user
needs to call `$bca->setReuseSession(true)`. This will make the library
not call logout() and not delete the cookie file upon the object
destruction. Note that by doing so, the user must take care of the
session themselves. They have to call logout() if they need to destroy
the session immediately. Otherwise, the account may get locked up for
10 minutes while waiting for the login session be expired. This change
doesn't affect users who don't care about reuse session.

Signed-off-by: Alviro Iskandar Setiawan <alviro.iskandar@gnuweeb.org>