KlikBCA: Introducing `setDumpHtmlErrorFile()` method
When an error happens while parsing the HTML page, it's hard to
conclude what is going wrong. Allow the user to set a dump error file
to dump the HTML page string into a file for manual investigation. This
method has to be called before calling other methods that perform an
HTTP request.

Signed-off-by: Ammar Faizi <ammarfaizi2@gnuweeb.org>