KlikBCA: Detect locked up pattern
Sometimes the account is locked up for 5 minutes when using this
library. It is normal because the cookies may get deleted without the
user navigating to the logout endpoint. The server only allows one
session at a time. When that happens, the server thinks the user
session is still active and rejects any other login attempts.

The login rejection will result in a message like this in a JavaScript
'alert()' function:

   "Anda dapat melakukan login kembali setelah 5 menit."

Report such a situation to the user via the error message string.
While in there, simplify the error handling code. Reduce code
duplication. Do the extra error construction in setError() instead of
in the callers.

Signed-off-by: Ammar Faizi <ammarfaizi2@gnuweeb.org>